Move config.py and config.yaml to root directory
